requirements.txt
py2neo==4.3.0
Flask==1.1.1
tensorlayer==2.2.2
jieba==0.39
tensorflow==1.13.1
numpy==1.18.1
pandas==1.0.1
zhon==1.1.5
pyahocorasick==1.4.0
gensim==3.8.0
scikit_learn==0.23.2
